project1/assets/scripts/requirements.txt
blis==0.2.4
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
construct==2.5.3
cymem==2.0.2
defusedxml==0.5.0
diff-match-patch==20181111
en-core-web-sm==2.1.0
et-xmlfile==1.0.1
future==0.17.1
get==2018.11.19
google-api-core==1.8.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud==0.34.0
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.8
grpcio==1.18.0
haystack==0.42
httplib2==0.12.1
idna==2.8
jdcal==1.4
jsonschema==2.6.0
lxml==4.3.2
magickwand==0.2
murmurhash==1.0.2
numpy==1.16.2
oauthlib==3.0.1
odfpy==1.4.0
opencv-python==4.0.0.21
openpyxl==2.6.1
pefile==2018.8.8
Pillow==6.0.0
plac==0.9.6
post==2018.11.20
preshed==2.0.1
protobuf==3.6.1
psycopg2==2.7.7
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pysolr==3.8.1
pytesseract==0.2.6
python-ptrace==0.9.3
pytz==2018.9
PyYAML==3.13
query-string==2018.11.20
reportlab==3.5.13
request==2018.11.20
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.2.1
six==1.12.0
spacy==2.1.3
srsly==0.0.5
tablib==0.12.1
thinc==7.0.4
tqdm==4.31.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
Wand==0.5.0
wasabi==0.2.1
xlrd==1.2.0
xlwt==1.3.0
